@@ -75,21 +75,21 @@
     "@types/react-dom": "^17.0.3",
     "@types/react-redux": "^7.1.16",
     "@vue/preload-webpack-plugin": "^2.0.0",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.0",
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.1",
     "autoprefixer": "^10.2.5",
     "copy-webpack-plugin": "^8.1.1",
-    "css-loader": "^5.2.0",
+    "css-loader": "^5.2.1",
     "css-modules-typescript-loader": "^4.0.1",
-    "cssnano": "^4.1.11",
+    "cssnano": "^5.0.0",
     "enzyme": "^3.11.0",
     "file-loader": "^6.2.0",
-    "fork-ts-checker-webpack-plugin": "^6.2.0",
+    "fork-ts-checker-webpack-plugin": "^6.2.1",
     "html-loader": "^2.1.2",
     "html-webpack-plugin": "^5.3.1",
     "husky": "^6.0.0",
     "jest": "^26.6.3",
     "node-sass": "^5.0.0",
-    "postcss": "^8.2.9",
+    "postcss": "^8.2.10",
     "postcss-import": "^14.0.1",
     "postcss-loader": "^5.2.0",
     "redux-devtools": "^3.7.0",
@@ -101,11 +101,11 @@
     "ts-jest": "^26.5.4",
     "ts-loader": "^8.1.0",
     "ts-node": "^9.1.1",
-    "typescript": "^4.2.3",
-    "webpack": "^5.30.0",
+    "typescript": "^4.2.4",
+    "webpack": "^5.32.0",
     "webpack-cli": "^4.6.0",
     "webpack-dev-server": "^3.11.2",
-    "workbox-webpack-plugin": "^6.1.2"
+    "workbox-webpack-plugin": "^6.1.5"
   },
   "optionalDependencies": {
     "@types/convert-excel-to-json": "^1.7.0",
@@ -114,12 +114,12 @@
     "@types/sharp": "^0.28.0",
     "convert-excel-to-json": "^1.7.0",
     "csv-parse": "^4.15.3",
-    "eslint": "^7.23.0",
+    "eslint": "^7.24.0",
     "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-prettier": "^8.1.0",
+    "eslint-config-prettier": "^8.2.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.23.1",
+    "eslint-plugin-react": "^7.23.2",
     "eslint-plugin-react-hooks": "^4.2.0",
     "fontmin": "^0.9.8",
     "http-server": "^0.12.3",
