@@ -14,6 +14,8 @@ import Window from './Window'
 import {
   UPDATE_SETTINGS_AND_INIT,
   SCREEN_PREF,
+  SWITCH_MULTIPLAYER_MODE,
+  CONNECT_TO_ID,
 } from '../../constants/ActionTypes'
 import { I18nContext, upper1st } from '../../i18n/I18nContext'
 import { preSettings } from '../../constants/preSettings'
@@ -23,8 +25,10 @@ import {
   defaultSettings,
 } from '../../constants/defaultSettings'
 import { hasOwnProperty } from '../../utils/typeHelpers'
-import { initPeer } from '../../webrtc/peer'
-import { copiedDuration } from '../../constants/visuals'
+import {
+  copiedDuration,
+  shorterIdStartEndLength,
+} from '../../constants/visuals'
 
 const Pref = () => {
   const _ = useContext(I18nContext)
@@ -32,6 +36,14 @@ const Pref = () => {
 
   const copied = useRef<HTMLDivElement | null>(null)
   const copiedTimer = useRef<NodeJS.Timeout | null>(null)
+
+  const isMultiplayer = useAppSelector((state) => state.multiplayer.on)
+  const yourId = useAppSelector((state) => state.multiplayer.yourId)
+  const multiplayerStatus = useAppSelector((state) => state.multiplayer.status)
+
+  const opponentIdInStore = useAppSelector(
+    (state) => state.multiplayer.opponentId,
+  )
 
   const [formFields, setFormFields] = useState({
     playerName: useAppSelector((state) => state.settings.playerName),
@@ -51,9 +63,7 @@ const Pref = () => {
 
     cardsInHand: useAppSelector((state) => state.settings.cardsInHand),
 
-    multiplayerMode: useAppSelector((state) => state.multiplayer.mode),
-    yourId: useAppSelector((state) => state.multiplayer.yourId),
-    opponentId: useAppSelector((state) => state.multiplayer.opponentId),
+    opponentId: opponentIdInStore,
   })
 
   const checkPreset = (o: typeof formFields) => {
@@ -94,9 +104,6 @@ const Pref = () => {
       winTower,
       winResource,
       cardsInHand,
-      // multiplayerMode,
-      // yourId,
-      // opponentId,
     } = formFields
 
     const payload = {
@@ -158,14 +165,27 @@ const Pref = () => {
     }
 
     setPreset(checkPreset(formFields))
+  }, [
+    formFields.tower,
+    formFields.wall,
+    formFields.bricks,
+    formFields.gems,
+    formFields.recruits,
+    formFields.brickProd,
+    formFields.gemProd,
+    formFields.recruitProd,
+    formFields.winTower,
+    formFields.winResource,
+    formFields.cardsInHand,
+  ])
 
-    if (formFields.multiplayerMode) {
-      const peer = initPeer()
-      peer.on('open', (id) => {
-        alert('My peer ID is: ' + id)
-      })
-    }
-  }, [formFields])
+  useEffect(() => {
+    setFormFields((prev) =>
+      produce(prev, (draft) => {
+        draft.opponentId = opponentIdInStore
+      }),
+    )
+  }, [opponentIdInStore])
 
   const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
     if (e.target.type === 'number') {
@@ -191,9 +211,6 @@ const Pref = () => {
             case 'opponentName':
             case 'opponentId':
               draft[name] = value
-              break
-            case 'multiplayerMode':
-              draft[name] = checked
               break
             default:
               draft[name] = parseInt(value, 10)
@@ -230,6 +247,65 @@ const Pref = () => {
     }
   }
 
+  let notification: string = ''
+  switch (multiplayerStatus) {
+    case 'connecting_net':
+      notification = _.i18n('Connecting to the network ⌛')
+      break
+
+    case 'connected_net':
+      notification = _.i18n('Connected to the network ✔️')
+      break
+
+    case 'connecting_id': {
+      const id = opponentIdInStore
+      const shorterId = `${id.substring(
+        0,
+        shorterIdStartEndLength,
+      )}...${id.substring(id.length - shorterIdStartEndLength)}`
+      notification = _.i18n('Connecting to ID %s ⌛').replace('%s', shorterId)
+      break
+    }
+
+    case 'connected_id': {
+      const id = opponentIdInStore
+      const shorterId = `${id.substring(
+        0,
+        shorterIdStartEndLength,
+      )}...${id.substring(id.length - shorterIdStartEndLength)}`
+      notification = _.i18n("Connected to ID %s ✔️ You're the host 🏠").replace(
+        '%s',
+        shorterId,
+      )
+      break
+    }
+
+    case 'connected_by': {
+      const id = opponentIdInStore
+      const shorterId = `${id.substring(
+        0,
+        shorterIdStartEndLength,
+      )}...${id.substring(id.length - shorterIdStartEndLength)}`
+      notification = _.i18n(
+        "Connected by ID %s ✔️ You're the guest 💼",
+      ).replace('%s', shorterId)
+      break
+    }
+
+    case 'failed':
+      notification = _.i18n('Connection failed ❌')
+      break
+
+    case 'disconnected':
+      notification = _.i18n('Disconnected 🔌')
+      break
+
+    default:
+      break
+  }
+
+  const isGuest = isMultiplayer && multiplayerStatus === 'connected_by'
+
   return (
     <Window ScreenActionType={SCREEN_PREF}>
       <h3>{_.i18n('Preferences')}</h3>
@@ -262,6 +338,7 @@ const Pref = () => {
             type="text"
             name="opponentName"
             id="opponentName"
+            disabled={isGuest}
             value={formFields.opponentName}
             onChange={handleChange}
             onFocus={(e) => {
@@ -281,7 +358,8 @@ const Pref = () => {
         <select
           name="tavern"
           id="tavern"
-          value={preset}
+          value={isGuest ? -2 : preset}
+          disabled={isGuest}
           onChange={(e) => {
             usePreset(parseInt(e.target.value, 10))
           }}
@@ -306,6 +384,7 @@ const Pref = () => {
             name="tower"
             id="tower"
             min="1"
+            disabled={isGuest}
             value={formFields.tower}
             onChange={handleChange}
           />
@@ -317,6 +396,7 @@ const Pref = () => {
             name="wall"
             id="wall"
             min="0"
+            disabled={isGuest}
             value={formFields.wall}
             onChange={handleChange}
           />
@@ -331,6 +411,7 @@ const Pref = () => {
             name="bricks"
             id="bricks"
             min="0"
+            disabled={isGuest}
             value={formFields.bricks}
             onChange={handleChange}
           />
@@ -342,6 +423,7 @@ const Pref = () => {
             name="brickProd"
             id="brickProd"
             min="1"
+            disabled={isGuest}
             value={formFields.brickProd}
             onChange={handleChange}
           />
@@ -356,6 +438,7 @@ const Pref = () => {
             name="gems"
             id="gems"
             min="0"
+            disabled={isGuest}
             value={formFields.gems}
             onChange={handleChange}
           />
@@ -367,6 +450,7 @@ const Pref = () => {
             name="gemProd"
             id="gemProd"
             min="1"
+            disabled={isGuest}
             value={formFields.gemProd}
             onChange={handleChange}
           />
@@ -381,6 +465,7 @@ const Pref = () => {
             name="recruits"
             id="recruits"
             min="0"
+            disabled={isGuest}
             value={formFields.recruits}
             onChange={handleChange}
           />
@@ -392,6 +477,7 @@ const Pref = () => {
             name="recruitProd"
             id="recruitProd"
             min="1"
+            disabled={isGuest}
             value={formFields.recruitProd}
             onChange={handleChange}
           />
@@ -407,6 +493,7 @@ const Pref = () => {
             name="winTower"
             id="winTower"
             min={(formFields.tower + 1).toString(10)}
+            disabled={isGuest}
             value={formFields.winTower}
             onChange={handleChange}
           />
@@ -424,6 +511,7 @@ const Pref = () => {
                 formFields.recruits + formFields.recruitProd,
               ) + 1
             ).toString(10)}
+            disabled={isGuest}
             value={formFields.winResource}
             onChange={handleChange}
           />
@@ -439,6 +527,7 @@ const Pref = () => {
           id="cardsInHand"
           min="0"
           max="15"
+          disabled={isGuest}
           value={formFields.cardsInHand}
           onChange={handleChange}
         />
@@ -455,21 +544,23 @@ const Pref = () => {
         <label>
           <input
             type="checkbox"
-            name="multiplayerMode"
-            id="multiplayerMode"
-            checked={formFields.multiplayerMode}
-            onChange={handleChange}
+            name="isMultiplayer"
+            id="isMultiplayer"
+            checked={isMultiplayer}
+            onChange={(e) => {
+              dispatch({
+                type: SWITCH_MULTIPLAYER_MODE,
+                on: e.target.checked,
+              })
+            }}
           />
           <span>
-            {_.i18n('Multiplayer Mode')}
+            {_.i18n('Multiplayer')}
             {_.i18n(': ')}
-            {_.i18n('off').toUpperCase()}
+            {(isMultiplayer ? _.i18n('on') : _.i18n('off')).toUpperCase()}
           </span>
         </label>
-        <span id="multiplayerNotification">
-          {_.i18n('Connecting to ID %s ⌛')}
-        </span>
-        {/* 'Connected to ID %s ✔️' */}
+        <span id="multiplayerNotification">{notification}</span>
       </h4>
       <div className="multiplayer">
         <label>
@@ -481,7 +572,7 @@ const Pref = () => {
             type="text"
             name="yourId"
             id="yourId"
-            value={formFields.yourId}
+            value={yourId}
             readOnly
             onClick={(e) => {
               const target = e.target as HTMLInputElement
@@ -514,6 +605,13 @@ const Pref = () => {
             {_.i18n(': ')}
           </span>
           <input
+            disabled={
+              !isMultiplayer ||
+              multiplayerStatus === 'connecting_net' ||
+              multiplayerStatus === 'connecting_id' ||
+              multiplayerStatus === 'connected_id' ||
+              multiplayerStatus === 'connected_by'
+            }
             type="text"
             name="opponentId"
             id="opponentId"
@@ -521,31 +619,51 @@ const Pref = () => {
             onChange={handleChange}
           />
         </label>
-        <button className="highlight" onClick={() => {}}>
+        <button
+          disabled={
+            !isMultiplayer ||
+            multiplayerStatus === 'connecting_net' ||
+            multiplayerStatus === 'connecting_id' ||
+            multiplayerStatus === 'connected_id' ||
+            multiplayerStatus === 'connected_by'
+          }
+          className="highlight"
+          onClick={(e) => {
+            dispatch({
+              type: CONNECT_TO_ID,
+              id: formFields.opponentId,
+            })
+          }}
+        >
           {_.i18n('Connect')}
         </button>
       </div>
 
       <div className="button-wrapper">
         <button
+          disabled={isGuest}
           onClick={() => {
             const { start, win, cardsInHand } = defaultSettings
-            setFormFields(({ multiplayerMode, yourId, opponentId }) => ({
+            setFormFields(({ opponentId }) => ({
               playerName: defaultPlayerName,
               opponentName: defaultOpponentName,
               ...start,
               winTower: win.tower,
               winResource: win.resource,
               cardsInHand: cardsInHand,
-              multiplayerMode,
-              yourId,
-              opponentId,
+              isMultiplayer, // unchanged
+              yourId, // unchanged
+              opponentId, // unchanged
             }))
           }}
         >
           {_.i18n('Reset')}
         </button>
-        <button className="warning" onClick={applyAndNewGame}>
+        <button
+          disabled={isGuest}
+          className="warning"
+          onClick={applyAndNewGame}
+        >
           {_.i18n('Apply & New Game')}
         </button>
       </div>
