@@ -3,21 +3,23 @@ import {
   SWITCH_MULTIPLAYER_MODE_MAIN,
   SET_YOUR_ID,
   SET_OPPONENT_ID,
+  MULTIPLAYER_STATUS,
 } from '../constants/ActionTypes'
 import { RootActionType } from '../types/actionObj'
 import { MultiplayerStateType } from '../types/state'
 
 const defaultMultiplayerState: MultiplayerStateType = {
-  mode: false,
+  on: false,
   yourId: '',
   opponentId: '',
+  status: 'disconnected',
 }
 
 const multiplayer = produce(
   (draft: MultiplayerStateType, action: RootActionType) => {
     switch (action.type) {
       case SWITCH_MULTIPLAYER_MODE_MAIN: {
-        draft.mode = action.on
+        draft.on = action.on
         break
       }
       case SET_YOUR_ID: {
@@ -26,6 +28,10 @@ const multiplayer = produce(
       }
       case SET_OPPONENT_ID: {
         draft.opponentId = action.id
+        break
+      }
+      case MULTIPLAYER_STATUS: {
+        draft.status = action.status
         break
       }
     }
